// pronti-frontend-main/front/src/app/constantes/lista-valores/lista-valores.enum.ts
export enum ListaValoresEnum {
  generoLibro = 'GL1',
  lugarNacimiento = 'LN1',
  tipoMedioContacto = 'TMC1',
  ciudadAgencia = 'CA1',
  tipoEmpresa = 'TE1',
  marca = 'MAA1',
  modelo = 'MOA1',
  color = 'COA1',
  tipoCuenta = 'TCB1',
  claseContribuyente = 'CLC1',
  cuentaContable = 'TCC1',
  tipoDocumentoIdentidad = 'TD1',
  paisUsuario = 'PU1',
  provinciaUsuario = 'PVU1',
  ciudadUsuario = 'CU1',
  parentesco = 'PT1',
  estadoCivil = 'EC1',
  grupoSanguineo = 'GS1',
  nivelEstudios = 'NET1',
  profesion = 'PRT1',
  codigoSectorial = 'CDS1',
  tipoDiscapacidad = 'TDC1',
  cargoE = 'CGE1',
  tipoContratoE = 'TCE1',
  ubicacionFisica = 'UBF1',
  tipoPrestamoE = 'TPE1',
  tipoCuentaBancaria = 'TCB1',
  tipoDocumento = 'TDD1',
  motivoSalida = 'MST1',
  medioDeCaptacion = 'MCC1',
  terminosPago = 'TMP1'
}